+++ src/plonemeeting/portal/core/testing.py
@@ -27,10 +27,16 @@ class PlonemeetingPortalCoreLayer(PloneSandboxLayer):
         # Load any other ZCML that is required for your tests.
         # The z3c.autoinclude feature is disabled in the Plone fixture base
         # layer.
+        import imio.omnia.assistant
+        import imio.omnia.core
+        import imio.omnia.tinymce
         import plone.restapi
 
         self.loadZCML(package=plone.restapi)
         self.loadZCML(package=collective.documentgenerator)
+        self.loadZCML(package=imio.omnia.core)
+        self.loadZCML(package=imio.omnia.assistant)
+        self.loadZCML(package=imio.omnia.tinymce)
         self.loadZCML(package=plonemeeting.portal.core)
 
         # Patch collective.fingerpointing