+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-2.4.2 (unreleased)
+2.4.2 (2026-06-15)
 ------------------
 
 - Send the ``Last-Modified`` header again on the ``@@custom_colors.css``